Rename csp.server.ts to csp-utils.server.ts
